requirements.txt
fastapi
uvicorn[standard]
sqlalchemy>=2.0
psycopg2-binary
alembic
python-dotenv
pydantic
pydantic[email]
pytest
httpx
asyncpg
passlib[bcrypt]
python-jose[cryptography]
